refactor: remove fetch-article on service
